Move param definitions to util module
Move param definitions to util module, instead to a resource
module.
